+++ components/dashboard/src/Menu.tsx
@@ -51,6 +51,9 @@ export default function Menu() {
     const showTeamsUI = user?.rolesOrPermissions?.includes('teams-and-projects') || window.location.hostname.endsWith('gitpod-dev.com') || window.location.hostname.endsWith('gitpod-io-dev.com');
     const team = getCurrentTeam(location, teams);
 
+    // Hide most of the top menu when in a full-page form.
+    const isMinimalUI = ['/new', '/teams/new'].includes(location.pathname);
+
     const [ teamMembers, setTeamMembers ] = useState<Record<string, TeamMemberInfo[]>>({});
     useEffect(() => {
         if (!showTeamsUI || !teams) {
@@ -161,7 +164,7 @@ export default function Menu() {
                                 <span className="flex-1 font-semibold">New Team</span>
                                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 14 14" className="w-3.5"><path fill="currentColor" fill-rule="evenodd" d="M7 0a1 1 0 011 1v5h5a1 1 0 110 2H8v5a1 1 0 11-2 0V8H1a1 1 0 010-2h5V1a1 1 0 011-1z" clip-rule="evenodd" /></svg>
                             </div>,
-                            onClick: () => history.push("/new-team"),
+                            onClick: () => history.push("/teams/new"),
                         }
                     ]}>
                         <div className="flex h-full p-2 mt-0.5">
@@ -179,13 +182,13 @@ export default function Menu() {
     }
 
     return <>
-        <header className="lg:px-28 px-10 flex flex-col pt-4 space-y-4">
-            <div className="flex">
+        <header className={`lg:px-28 px-10 flex flex-col pt-4 space-y-4 ${isMinimalUI ? 'pb-4' : ''}`}>
+            <div className="flex h-10">
                 <div className="flex justify-between items-center pr-3">
                     <Link to="/">
                         <img src={gitpodIcon} className="h-6" />
                     </Link>
-                    <div className="ml-2 text-base">
+                    {!isMinimalUI && <div className="ml-2 text-base">
                         {showTeamsUI
                             ? renderTeamMenu()
                             : <nav className="flex-1">
@@ -197,13 +200,13 @@ export default function Menu() {
                                 </ul>
                             </nav>
                         }
-                    </div>
+                    </div>}
                 </div>
                 <div className="flex flex-1 items-center w-auto" id="menu">
                     <nav className="flex-1">
                         <ul className="flex flex-1 items-center justify-between text-base text-gray-700 space-x-2">
                             <li className="flex-1"></li>
-                            {rightMenu.map(entry => <li key={entry.title}>
+                            {!isMinimalUI && rightMenu.map(entry => <li key={entry.title}>
                                 <PillMenuItem name={entry.title} selected={isSelected(entry, location)} link={entry.link}/>
                             </li>)}
                         </ul>
@@ -230,7 +233,7 @@ export default function Menu() {
                     </div>
                 </div>
             </div>
-            {showTeamsUI && <div className="flex">
+            {!isMinimalUI && showTeamsUI && <div className="flex">
                 {leftMenu.map(entry => <TabMenuItem name={entry.title} selected={isSelected(entry, location)} link={entry.link}/>)}
             </div>}
         </header>