+++ components/dashboard/src/teams/NewTeam.tsx
@@ -27,7 +27,7 @@ export default function () {
             setCreationError(error);
         }
     }
-    return <div className="flex flex-col w-96 mt-16 mx-auto items-center">
+    return <div className="flex flex-col w-96 mt-24 mx-auto items-center">
         <h1>New Team</h1>
         <p className="text-gray-500 text-center text-base">Teams allow you to <strong>group multiple projects</strong>, <strong>collaborate with others</strong>, <strong>manage subscriptions</strong> with one centralized billing, and more. <a className="learn-more" href="https://www.gitpod.io/docs/teams/">Learn more</a></p>
         <form className="mt-16 w-full" onSubmit={createTeam}>